fix(seeders): register ChargingStationSeeder and fix coordinate types
- Add ChargingStationSeeder to DatabaseSeeder call stack
- Change latitude/longitude from integer to double for proper coordinates
- Add 3 sample charging stations (Alpha, Beta, Gamma) with realistic data

+++ database/migrations/2026_03_10_122321_create_charging_stations_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('charging_stations', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('latitude');
-            $table->integer('longitude');
+            $table->double('latitude');
+            $table->double('longitude');
             $table->string('connector_type');
             $table->boolean('is_available');
             $table->timestamps();
